requirements.txt
aiosqlite==0.19.0
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
fastapi==0.106.0
greenlet==3.0.3
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
mypy==1.8.0
mypy-extensions==1.0.0
pydantic==2.5.3
pydantic_core==2.14.6
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlalchemy-stubs==0.4
starlette==0.27.0
typing_extensions==4.9.0

python-dotenv~=1.0.0
requests~=2.31.0